ISSUE 8314 allow production version of jarvis work with the latest XML smart lib.
XML smart now uses wantarray so you have to be careful of the context when calling the
content method on XML. It would seem that inside the map function the context gets set
to array somehow. By explicitly setting this to go in a scalar variable it seems to set
the context correctly.

--HG--
branch : SUSTAINING_5_4_10